Fixed GameWordInput
- Now correctly shows the last letter of the last word given
