+++ app/games/components/GameLoader.tsx
@@ -1,5 +1,16 @@
-import { useParam, useQuery } from "blitz"
-import { Center, Wrap, WrapItem } from "@chakra-ui/react"
+import { useMutation, useParam, useQuery, useRouter } from "blitz"
+import {
+  Button,
+  Center,
+  ModalBody,
+  ModalCloseButton,
+  ModalContent,
+  ModalFooter,
+  ModalHeader,
+  ModalOverlay,
+  Wrap,
+  WrapItem,
+} from "@chakra-ui/react"
 import { Suspense, useEffect } from "react"
 import { PlayerGameCard } from "../../players/components/PlayerGameCard"
 import { LoadingSpinner } from "../../core/components/LoadingSpinner"
@@ -7,6 +18,8 @@ import { useChannel, useEvent } from "@harelpls/use-pusher"
 import getWords from "../../words/queries/getWords"
 import { Game, Player, Word } from "db"
 import getGameWithPlayers from "../queries/getGameWithPlayers"
+import { UnCloseableModal } from "../../core/components/UnCloseableModal"
+import logoutMutation from "../../auth/mutations/logout"
 
 type GameWithPlayers = Game & {
   winner: Player | null
@@ -14,11 +27,23 @@ type GameWithPlayers = Game & {
 }
 
 const GameUI = ({ gameId }) => {
+  const router = useRouter()
+  const [logout] = useMutation(logoutMutation)
   const [game, { refetch }] = useQuery(getGameWithPlayers, { id: gameId })
   const channel = useChannel(gameId)
 
+  const redirectToHome = async () => {
+    await logout()
+    await router.replace("/")
+  }
+
   useEvent(channel, "turn-taken", async (data) => {
     await refetch()
+  })
+
+  useEvent(channel, "game-finished", async (data) => {
+    await refetch()
+    game.finished = true
   })
 
   useEffect(() => {}, [gameId])
@@ -27,6 +52,21 @@ const GameUI = ({ gameId }) => {
       <Center>
         <GamePlayerList game={game} />
       </Center>
+
+      <UnCloseableModal isOpen={game.finished} onClose={() => {}}>
+        <ModalOverlay />
+        <ModalContent>
+          <ModalHeader>Game Finished</ModalHeader>
+          <ModalCloseButton />
+          <ModalBody>Congratulations {game.winner?.name}</ModalBody>
+
+          <ModalFooter>
+            <Button variant="ghost" onClick={redirectToHome}>
+              Go Home
+            </Button>
+          </ModalFooter>
+        </ModalContent>
+      </UnCloseableModal>
       {/*<GameWordList gameId={gameId!} />*/}
     </>
   )
