+++ app/games/components/forms/GameWordForm.tsx
@@ -2,6 +2,7 @@ import { Form, FormProps } from "app/core/components/Form"
 import { z } from "zod"
 import { Input } from "@chakra-ui/react"
 import { useController } from "react-hook-form"
+import { useState } from "react"
 
 export { FORM_ERROR } from "app/core/components/Form"
 
@@ -18,8 +19,10 @@ const GameWordInput = ({
   isActivePlayer: boolean
   lastWord?: string | null
 }) => {
+  const [word, setWord] = useState<string>(lastWord?.charAt(lastWord?.length - 1) || "")
+
   const {
-    field,
+    field: { value, onChange, ...fieldprops },
     formState: { isSubmitting },
   } = useController({
     name: "word",
@@ -27,8 +30,21 @@ const GameWordInput = ({
     defaultValue: lastWord?.charAt(lastWord?.length - 1),
   })
 
+  const onWordChange = (event) => {
+    const value = event.target.value
+    setWord(value)
+    onChange(value)
+  }
+
   return (
-    <Input id={`field-name`} isDisabled={isSubmitting || !isActivePlayer} {...field} {...props} />
+    <Input
+      id={`field-word`}
+      onChange={onWordChange}
+      value={word}
+      isDisabled={isSubmitting || !isActivePlayer}
+      {...fieldprops}
+      {...props}
+    />
   )
 }
